requirements.txt
python-decouple
ncclient
schedule
psycopg2

